Reminders/Alerts now work
Dates are now saved in the task container. These are used when being sent to the AlertCenter class to be added to the system alerts. If reminders are switched off then the alert will not be created. Edit page still needs to be able to edit the tasks properly. Probably best to do it through the TaskManager class.